requirements.txt
# Core
python-dotenv==1.0.1
pandas==2.2.0
numpy==1.26.0
scikit-learn==1.4.0
google-adk>=1.17.0,<1.20.0

# Data Drift
evidently==0.4.19
scipy==1.12.0

# Visualization
matplotlib==3.8.2
seaborn==0.13.0

# ML
xgboost==2.0.3
lightgbm==4.3.0

# Testing
pytest==8.0.0
pytest-cov==4.1.0

# Kaggle API
kaggle==1.6.0

# Monitoring
mlflow>=2.10.0,<3.0.0
dvc==3.40.0

# Utils
pyyaml>=6.0.2,<7.0.0
joblib==1.3.2